+++ android/app/src/main/AndroidManifest.xml
@@ -12,9 +12,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Ohmycarset"
         tools:targetApi="31">
-        <activity android:name=".MainActivity"/>
+        <activity android:name=".ui.activity.MainActivity"/>
         <activity
-            android:name=".OnBoardingActivity"
+            android:name=".ui.activity.OnBoardingActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
